major rework
