+++ package.json
@@ -15,16 +15,24 @@
     "url": "https://github.com/RyanByers1023/Commerce-Bank-Project/issues"
   },
   "homepage": "https://github.com/RyanByers1023/Commerce-Bank-Project#readme",
-  "description": "",
   "dependencies": {
     "@tailwindcss/cli": "^4.0.7",
     "@tailwindcss/vite": "^4.0.7",
+    "bcrypt": "^5.1.1",
+    "cors": "^2.8.5",
+    "express": "^5.1.0",
+    "express-mysql-session": "^3.0.3",
+    "express-session": "^1.18.1",
+    "mysql2": "^3.14.1",
     "node-cron": "^3.0.3",
-    "node-fetch": "^3.3.2"
+    "node-fetch": "^3.3.2",
+    "uuid": "^11.1.0"
   },
   "devDependencies": {
     "autoprefixer": "^10.4.20",
     "postcss": "^8.5.3",
     "tailwindcss": "^4.0.7"
-  }
+  },
+  "keywords": [],
+  "description": ""
 }
