+++ src/client/css/input.css
@@ -1,0 +1,1 @@
+@import "../../../node_modules/tailwindcss";
