--- src/navbar.html
@@ -1,23 +1,0 @@
-<nav class="bg-[#3A4F41] p-4 fixed top-0 w-full shadow-md">
-    <div class="max-w-5xl mx-auto flex justify-between text-white">
-        <!-- Left Side: Home & About -->
-        <div class="flex space-x-6">
-            <a href="index.html" class="text-lg font-semibold hover:text-[#2EF090]">Home</a>
-            <a href="about.html" class="text-lg text-white font-semibold hover:text-[#2EF090]">About Us</a>
-        </div>
-
-        <!-- Right Side: Stock Simulator & Login/Register Dropdown -->
-        <div class="flex space-x-6">
-            <a href="simulator.html" class="text-lg font-semibold hover:text-[#2EF090]">Stock Simulator</a>
-
-            <!-- Dropdown Menu -->
-            <div class="relative group">
-                <button class="text-lg font-semibold hover:text-[#2EF090]">Login/Register</button>
-                <div class="absolute hidden group-hover:block bg-[#4D705F] text-white mt-0.4 py-2 rounded-lg shadow-lg w-40">
-                    <a href="login.html" class="block px-4 py-2 hover:bg-[#2EF090] rounded-t-lg">Login</a>
-                    <a href="register.html" class="block px-4 py-2 hover:bg-[#2EF090] rounded-b-lg">Register</a>
-                </div>
-            </div>
-        </div>
-    </div>
-</nav>
